+++ csharp/rust/src/lib.rs
@@ -35,20 +35,32 @@ pub struct Client {
 /// The success callback needs to copy the given data synchronously, since it will be dropped by Rust once the callback returns.
 /// The callback should be offloaded to a separate thread in order not to exhaust the client's thread pool.
 ///
-/// `index` is a baton-pass back to the caller language to uniquely identify the promise.
-/// `message` is the value returned by the command. The 'message' is managed by Rust and is freed when the callback returns control back to the caller.
-pub type SuccessCallback = unsafe extern "C" fn(usize, *const ResponseValue) -> ();
+/// # Arguments
+/// * `index` is a baton-pass back to the caller language to uniquely identify the promise.
+/// * `message` is the value returned by the command. The 'message' is managed by Rust and is freed when the callback returns control back to the caller.
+///
+/// # Safety
+/// * The callback must copy the pointer in a sync manner and return ASAP. Any further data processing should be done in another thread to avoid
+///   starving `tokio`'s thread pool.
+/// * The callee is responsible to free memory by calling [`free_respose`] with the given pointer once only.
+pub type SuccessCallback = unsafe extern "C-unwind" fn(usize, *const ResponseValue) -> ();
 
 /// Failure callback that is called when a command fails.
 ///
 /// The failure callback needs to copy the given string synchronously, since it will be dropped by Rust once the callback returns.
 /// The callback should be offloaded to a separate thread in order not to exhaust the client's thread pool.
 ///
+/// # Arguments
 /// * `index` is a baton-pass back to the caller language to uniquely identify the promise.
 /// * `error_message` is an UTF-8 string storing the error message returned by server for the failed command.
 ///   The `error_message` is managed by Rust and is freed when the callback returns control back to the caller.
 /// * `error_type` is the type of error returned by glide-core, depending on the [`RedisError`](redis::RedisError) returned.
-pub type FailureCallback = unsafe extern "C" fn(
+///
+/// # Safety
+/// * The callback must copy the data in a sync manner and return ASAP. Any further data processing should be done in another thread to avoid
+///   starving `tokio`'s thread pool.
+/// * The caller must free the memory allocated for [`error_message`] right after the call to avoid memory leak.
+pub type FailureCallback = unsafe extern "C-unwind" fn(
     index: usize,
     error_message: *const c_char,
     error_type: RequestErrorType,
@@ -60,6 +72,45 @@ struct CommandExecutionCore {
     failure_callback: FailureCallback,
 }
 
+/// # Safety
+/// Unsafe, becase calls to an FFI function. See the safety documentation of [`FailureCallback`].
+unsafe fn report_error(
+    failure_callback: FailureCallback,
+    callback_index: usize,
+    error_string: String,
+    error_type: RequestErrorType,
+) {
+    logger_core::log(logger_core::Level::Error, "ffi", &error_string);
+    let err_ptr = CString::into_raw(
+        CString::new(error_string).expect("Couldn't convert error message to CString"),
+    );
+    unsafe { failure_callback(callback_index, err_ptr, error_type) };
+    // free memory
+    _ = CString::from_raw(err_ptr);
+}
+
+/// Panic Guard as per <https://www.reddit.com/r/rust/comments/zg2xcu/comment/izi758v/>
+struct PanicGuard {
+    panicked: bool,
+    failure_callback: FailureCallback,
+    callback_index: usize,
+}
+
+impl Drop for PanicGuard {
+    fn drop(&mut self) {
+        if self.panicked {
+            unsafe {
+                report_error(
+                    self.failure_callback,
+                    self.callback_index,
+                    "Native function panicked".into(),
+                    RequestErrorType::Unspecified,
+                );
+            }
+        }
+    }
+}
+
 /// Creates a new client with the given configuration.
 /// The success callback needs to copy the given string synchronously, since it will be dropped by Rust once the callback returns.
 /// All callbacks should be offloaded to separate threads in order not to exhaust the client's thread pool.
@@ -67,13 +118,21 @@ struct CommandExecutionCore {
 /// # Safety
 ///
 /// * `config` must be a valid [`ConnectionConfig`] pointer. See the safety documentation of [`create_connection_request`].
+/// * `success_callback` and `failure_callback` must be valid pointers to the corresponding FFI functions.
+///   See the safety documentation of [`SuccessCallback`] and [`FailureCallback`].
 #[allow(rustdoc::private_intra_doc_links)]
 #[no_mangle]
-pub unsafe extern "C" fn create_client(
+pub unsafe extern "C-unwind" fn create_client(
     config: *const ConnectionConfig,
     success_callback: SuccessCallback,
     failure_callback: FailureCallback,
 ) {
+    let mut panic_guard = PanicGuard {
+        panicked: true,
+        failure_callback,
+        callback_index: 0,
+    };
+
     let request = unsafe { create_connection_request(config) };
     let runtime = Builder::new_multi_thread()
         .enable_all()
@@ -95,13 +154,19 @@ pub unsafe extern "C" fn create_client(
             unsafe { success_callback(0, client_ptr as *const ResponseValue) };
         }
         Err(err) => {
-            let err_ptr = CString::into_raw(
-                CString::new(err.to_string()).expect("Couldn't convert error message to CString"),
-            );
-            unsafe { failure_callback(0, err_ptr, RequestErrorType::Disconnect) };
-            _ = CString::from_raw(err_ptr);
+            unsafe {
+                report_error(
+                    failure_callback,
+                    0,
+                    err.to_string(),
+                    RequestErrorType::Disconnect,
+                )
+            };
         }
     }
+
+    panic_guard.panicked = false;
+    drop(panic_guard);
 }
 
 /// Closes the given client, deallocating it from the heap.
@@ -111,7 +176,7 @@ pub unsafe extern "C" fn create_client(
 /// # Safety
 ///
 /// * `client_ptr` must not be `null`.
-/// * `client_ptr` must be able to be safely casted to a valid [`Box<Client>`] via [`Box::from_raw`]. See the safety documentation of [`Box::from_raw`].
+/// * `client_ptr` must be able to be safely casted to a valid [`Arc<Client>`] via [`Arc::from_raw`]. See the safety documentation of [`Arc::from_raw`].
 #[no_mangle]
 pub extern "C" fn close_client(client_ptr: *const c_void) {
     assert!(!client_ptr.is_null());
@@ -123,14 +188,15 @@ pub extern "C" fn close_client(client_ptr: *const c_void) {
 ///
 /// # Safety
 /// * `client_ptr` must not be `null`.
-/// * `client_ptr` must be able to be safely casted to a valid [`Box<Client>`] via [`Box::from_raw`]. See the safety documentation of [`Box::from_raw`].
+/// * `client_ptr` must be able to be safely casted to a valid [`Arc<Client>`] via [`Arc::from_raw`]. See the safety documentation of [`Arc::from_raw`].
 /// * This function should only be called should with a pointer created by [`create_client`], before [`close_client`] was called with the pointer.
+/// * Pointers to callbacks stored in [`Client`] should remain valid. See the safety documentation of [`SuccessCallback`] and [`FailureCallback`].
 /// * `cmd_ptr` must not be `null`.
 /// * `cmd_ptr` must be able to be safely casted to a valid [`CmdInfo`]. See the safety documentation of [`create_cmd`].
 /// * `route_info` could be `null`, but if it is not `null`, it must be a valid [`RouteInfo`] pointer. See the safety documentation of [`create_route`].
 #[allow(rustdoc::private_intra_doc_links)]
 #[no_mangle]
-pub unsafe extern "C" fn command(
+pub unsafe extern "C-unwind" fn command(
     client_ptr: *const c_void,
     callback_index: usize,
     cmd_ptr: *const CmdInfo,
@@ -143,16 +209,23 @@ pub unsafe extern "C" fn command(
     };
     let core = client.core.clone();
 
+    let mut panic_guard = PanicGuard {
+        panicked: true,
+        failure_callback: core.failure_callback,
+        callback_index,
+    };
+
     let cmd = match unsafe { create_cmd(cmd_ptr) } {
         Ok(cmd) => cmd,
         Err(err) => {
-            let err_ptr = CString::into_raw(
-                CString::new(err).expect("Couldn't convert error message to CString"),
-            );
             unsafe {
-                (core.failure_callback)(callback_index, err_ptr, RequestErrorType::ExecAbort)
-            };
-            _ = CString::from_raw(err_ptr);
+                report_error(
+                    core.failure_callback,
+                    callback_index,
+                    err,
+                    RequestErrorType::Unspecified,
+                );
+            }
             return;
         }
     };
@@ -160,22 +233,33 @@ pub unsafe extern "C" fn command(
     let route = unsafe { create_route(route_info, Some(&cmd)) };
 
     client.runtime.spawn(async move {
+        let mut panic_guard = PanicGuard {
+            panicked: true,
+            failure_callback: core.failure_callback,
+            callback_index,
+        };
+
         let result = core.client.clone().send_command(&cmd, route).await;
         match result {
             Ok(value) => {
                 let ptr = Box::into_raw(Box::new(ResponseValue::from_value(value)));
                 unsafe { (core.success_callback)(callback_index, ptr) };
             }
-            Err(err) => {
-                let err_ptr = CString::into_raw(
-                    CString::new(error_message(&err))
-                        .expect("Couldn't convert error message to CString"),
+            Err(err) => unsafe {
+                report_error(
+                    core.failure_callback,
+                    callback_index,
+                    error_message(&err),
+                    error_type(&err),
                 );
-                unsafe { (core.failure_callback)(callback_index, err_ptr, error_type(&err)) };
-                _ = CString::from_raw(err_ptr);
-            }
+            },
         };
+        panic_guard.panicked = false;
+        drop(panic_guard);
     });
+
+    panic_guard.panicked = false;
+    drop(panic_guard);
 }
 
 /// Execute a batch.
@@ -202,16 +286,23 @@ pub unsafe extern "C" fn batch(
     };
     let core = client.core.clone();
 
+    let mut panic_guard = PanicGuard {
+        panicked: true,
+        failure_callback: core.failure_callback,
+        callback_index,
+    };
+
     let pipeline = match unsafe { create_pipeline(batch_ptr) } {
         Ok(pipeline) => pipeline,
         Err(err) => {
-            let err_ptr = CString::into_raw(
-                CString::new(err).expect("Couldn't convert error message to CString"),
-            );
             unsafe {
-                (core.failure_callback)(callback_index, err_ptr, RequestErrorType::ExecAbort)
-            };
-            _ = CString::from_raw(err_ptr);
+                report_error(
+                    core.failure_callback,
+                    callback_index,
+                    err,
+                    RequestErrorType::Unspecified,
+                );
+            }
             return;
         }
     };
@@ -220,6 +311,12 @@ pub unsafe extern "C" fn batch(
         unsafe { get_pipeline_options(options_ptr) };
 
     client.runtime.spawn(async move {
+        let mut panic_guard = PanicGuard {
+            panicked: true,
+            failure_callback: core.failure_callback,
+            callback_index,
+        };
+
         let result = if pipeline.is_atomic() {
             core.client
                 .clone()
@@ -242,16 +339,21 @@ pub unsafe extern "C" fn batch(
                 let ptr = Box::into_raw(Box::new(ResponseValue::from_value(value)));
                 unsafe { (core.success_callback)(callback_index, ptr) };
             }
-            Err(err) => {
-                let err_ptr = CString::into_raw(
-                    CString::new(error_message(&err))
-                        .expect("Couldn't convert error message to CString"),
+            Err(err) => unsafe {
+                report_error(
+                    core.failure_callback,
+                    callback_index,
+                    error_message(&err),
+                    error_type(&err),
                 );
-                unsafe { (core.failure_callback)(callback_index, err_ptr, error_type(&err)) };
-                _ = CString::from_raw(err_ptr);
-            }
+            },
         };
+        panic_guard.panicked = false;
+        drop(panic_guard);
     });
+
+    panic_guard.panicked = false;
+    drop(panic_guard);
 }
 
 /// Free the memory allocated for a [`ResponseValue`] and nested structure.
@@ -263,8 +365,7 @@ pub unsafe extern "C" fn batch(
 #[no_mangle]
 pub unsafe extern "C" fn free_respose(ptr: *mut ResponseValue) {
     unsafe {
-        let val = Box::leak(Box::from_raw(ptr));
-        val.free_memory();
+        Box::leak(Box::from_raw(ptr)).free_memory();
     }
 }